New pets worker and pets end point

+++ utils/generateNewWorker.js
@@ -3,9 +3,11 @@ const path = require('path');
 const requestTracker = require('./requestTracker');
 const workerTracker = require('./workerTracker');
 
-const WORKER_IDLE_TIME = 15*60*1000;
+const WORKER_IDLE_TIME = 1*60*1000;
 
-const getWorker = (workerName, currentTime) => {
+const getWorker = (workerName) => {
+  const currentTime = Date.now();
+  
   if (!workerTracker[workerName] || (currentTime - workerTracker[workerName].startedTime) > WORKER_IDLE_TIME) {
     workerTracker[workerName] = {startedTime: currentTime, worker: generateNewWorker(workerName)};
   }
@@ -13,21 +15,28 @@ const getWorker = (workerName, currentTime) => {
   return workerTracker[workerName].worker; 
 }
 
+const terminateWorker = (workerName, worker) => {
+  if (workerTracker[workerName]) {
+    console.log('Terminating worker:', workerName);
+    worker.terminate();
+    delete workerTracker[workerName];
+  }
+}
+
 const generateNewWorker = (workerName) => {
   console.log('Creating new worker:', workerName);
-  const worker = new Worker(path.join(__dirname, '../workers', workerName));
+  const worker = new Worker(path.join(__dirname, '../workers', 'pets'));
   worker.on('message', (data) => {
     const { response, requestId } = data;
     requestTracker[requestId](response);
     delete requestTracker[requestId];
   });
   worker.on('error', () => {
-    worker.terminate();
+    terminateWorker(workerName, worker);
   });
 
   setTimeout(() => { 
-    console.log('Terminating worker:', workerName);
-    worker.terminate()
+    terminateWorker(workerName, worker);
   },WORKER_IDLE_TIME);
   return worker;
 }
